Update tests for abstract types and include new failing test
This commit updates the test suite for abstract types to include
inline snapshots in order to easily introspect the expansion of
type conditions that is part of the current model.

This commit also adds a new (and failing) test to describe an improved
behavior where interfaces that exist within only a single serivce
can bypass being expanded at all.