PoC: Attempt APQ when fetching to downstream services from gateway.
Ignoring the fact that this breaks a number of test expectations which are
difficult to rejigger with the current test harness, this implementation
also has a number of concerns to be addressed in terms of complying with
existing expectations.  For example, `didReceiveResponse` can now called more
than once per downstream fetch, which might be unexpected by sub-classes.

To be honest, there's a fair amount of API consideration to be had before
I'd call this ready.